+++ cli/tests/pcluster/config/test_validators.py
@@ -694,7 +694,7 @@ def test_efs_validator(mocker, section_dict, expected_message):
             {"volume_type": "io1", "volume_size": 20, "volume_iops": 64001},
             "IOPS rate must be between 100 and 64000 when provisioning io1 volumes.",
         ),
-        ({"volume_type": "io1", "volume_size": 20, "volume_iops": 1001}, "IOPS to volume size ratio of .* is too hig"),
+        ({"volume_type": "io1", "volume_size": 20, "volume_iops": 1001}, "IOPS to volume size ratio of .* is too high"),
         ({"volume_type": "io2", "volume_size": 20, "volume_iops": 120}, None),
         (
             {"volume_type": "io2", "volume_size": 20, "volume_iops": 90},
@@ -704,7 +704,23 @@ def test_efs_validator(mocker, section_dict, expected_message):
             {"volume_type": "io2", "volume_size": 20, "volume_iops": 64001},
             "IOPS rate must be between 100 and 64000 when provisioning io2 volumes.",
         ),
-        ({"volume_type": "io2", "volume_size": 20, "volume_iops": 10001}, "IOPS to volume size ratio of .* is too hig"),
+        (
+            {"volume_type": "io2", "volume_size": 20, "volume_iops": 10001},
+            "IOPS to volume size ratio of .* is too high",
+        ),
+        ({"volume_type": "gp3", "volume_size": 20, "volume_iops": 3000}, None),
+        (
+            {"volume_type": "gp3", "volume_size": 20, "volume_iops": 2900},
+            "IOPS rate must be between 3000 and 16000 when provisioning gp3 volumes.",
+        ),
+        (
+            {"volume_type": "gp3", "volume_size": 20, "volume_iops": 16001},
+            "IOPS rate must be between 3000 and 16000 when provisioning gp3 volumes.",
+        ),
+        (
+            {"volume_type": "gp3", "volume_size": 20, "volume_iops": 10001},
+            "IOPS to volume size ratio of .* is too high",
+        ),
     ],
 )
 def test_raid_validators(mocker, section_dict, expected_message):
@@ -2437,6 +2453,9 @@ def test_fsx_ignored_parameters_validator(mocker, section_dict, expected_error):
         ({"volume_type": "gp2", "volume_size": 15}, None),
         ({"volume_type": "gp2", "volume_size": 0}, "The size of gp2 volumes must be at least 1 GiB"),
         ({"volume_type": "gp2", "volume_size": 16385}, "The size of gp2 volumes can not exceed 16384 GiB"),
+        ({"volume_type": "gp3", "volume_size": 15}, None),
+        ({"volume_type": "gp3", "volume_size": 0}, "The size of gp3 volumes must be at least 1 GiB"),
+        ({"volume_type": "gp3", "volume_size": 16385}, "The size of gp3 volumes can not exceed 16384 GiB"),
         ({"volume_type": "st1", "volume_size": 500}, None),
         ({"volume_type": "st1", "volume_size": 20}, "The size of st1 volumes must be at least 500 GiB"),
         ({"volume_type": "st1", "volume_size": 16385}, "The size of st1 volumes can not exceed 16384 GiB"),
@@ -2470,7 +2489,7 @@ def test_ebs_allowed_values_all_have_volume_size_bounds():
             {"volume_type": "io1", "volume_size": 20, "volume_iops": 64001},
             "IOPS rate must be between 100 and 64000 when provisioning io1 volumes.",
         ),
-        ({"volume_type": "io1", "volume_size": 20, "volume_iops": 1001}, "IOPS to volume size ratio of .* is too hig"),
+        ({"volume_type": "io1", "volume_size": 20, "volume_iops": 1001}, "IOPS to volume size ratio of .* is too high"),
         ({"volume_type": "io2", "volume_size": 20, "volume_iops": 120}, None),
         (
             {"volume_type": "io2", "volume_size": 20, "volume_iops": 90},
@@ -2480,7 +2499,23 @@ def test_ebs_allowed_values_all_have_volume_size_bounds():
             {"volume_type": "io2", "volume_size": 20, "volume_iops": 64001},
             "IOPS rate must be between 100 and 64000 when provisioning io2 volumes.",
         ),
-        ({"volume_type": "io2", "volume_size": 20, "volume_iops": 10001}, "IOPS to volume size ratio of .* is too hig"),
+        (
+            {"volume_type": "io2", "volume_size": 20, "volume_iops": 10001},
+            "IOPS to volume size ratio of .* is too high",
+        ),
+        ({"volume_type": "gp3", "volume_size": 20, "volume_iops": 3000}, None),
+        (
+            {"volume_type": "gp3", "volume_size": 20, "volume_iops": 2900},
+            "IOPS rate must be between 3000 and 16000 when provisioning gp3 volumes.",
+        ),
+        (
+            {"volume_type": "gp3", "volume_size": 20, "volume_iops": 16001},
+            "IOPS rate must be between 3000 and 16000 when provisioning gp3 volumes.",
+        ),
+        (
+            {"volume_type": "gp3", "volume_size": 20, "volume_iops": 10001},
+            "IOPS to volume size ratio of .* is too high",
+        ),
     ],
 )
 def test_ebs_volume_iops_validator(mocker, section_dict, expected_message):
@@ -2711,3 +2746,28 @@ def test_efa_os_arch_validator(mocker, cluster_dict, architecture, expected_erro
         assert_that(errors[0]).matches(expected_error)
     else:
         assert_that(errors).is_empty()
+
+
+@pytest.mark.parametrize(
+    "section_dict, expected_message",
+    [
+        ({"volume_type": "gp3", "volume_throughput": 125}, None),
+        (
+            {"volume_type": "gp3", "volume_throughput": 100},
+            "Throughput must be between 125 MB/s and 1000 MB/s when provisioning gp3 volumes.",
+        ),
+        (
+            {"volume_type": "gp3", "volume_throughput": 1001},
+            "Throughput must be between 125 MB/s and 1000 MB/s when provisioning gp3 volumes.",
+        ),
+        ({"volume_type": "gp3", "volume_throughput": 125, "volume_iops": 3000}, None),
+        (
+            {"volume_type": "gp3", "volume_throughput": 760, "volume_iops": 3000},
+            "Throughput to IOPS ratio of .* is too high",
+        ),
+        ({"volume_type": "gp3", "volume_throughput": 760, "volume_iops": 10000}, None),
+    ],
+)
+def test_ebs_volume_throughput_validator(mocker, section_dict, expected_message):
+    config_parser_dict = {"cluster default": {"ebs_settings": "default"}, "ebs default": section_dict}
+    utils.assert_param_validator(mocker, config_parser_dict, expected_message)